Moved trace implementation to heimdall-core and removed StackTrace from middleware spec
