+++ heimdall-gateway/src/main/java/br/com/conductor/heimdall/gateway/filter/CustomHostRoutingFilter.java
@@ -21,7 +21,7 @@ package br.com.conductor.heimdall.gateway.filter;
 
 import br.com.conductor.heimdall.core.util.Constants;
 import br.com.conductor.heimdall.gateway.failsafe.CircuitBreakerManager;
-import br.com.conductor.heimdall.gateway.trace.FilterDetail;
+import br.com.conductor.heimdall.core.trace.FilterDetail;
 import br.com.conductor.heimdall.gateway.trace.TraceContextHolder;
 import com.netflix.zuul.context.RequestContext;
 import lombok.extern.slf4j.Slf4j;
