+++ heimdall-gateway/src/main/java/br/com/conductor/heimdall/gateway/filter/CustomSendResponseFilter.java
@@ -26,7 +26,7 @@ import org.springframework.cloud.netflix.zuul.filters.post.SendResponseFilter;
 import com.netflix.zuul.context.RequestContext;
 
 import br.com.conductor.heimdall.core.util.Constants;
-import br.com.conductor.heimdall.gateway.trace.FilterDetail;
+import br.com.conductor.heimdall.core.trace.FilterDetail;
 import br.com.conductor.heimdall.gateway.trace.TraceContextHolder;
 import lombok.extern.slf4j.Slf4j;
