+++ heimdall-gateway/src/main/java/br/com/conductor/heimdall/gateway/filter/ScopesFilter.java
@@ -22,8 +22,7 @@ package br.com.conductor.heimdall.gateway.filter;
 import br.com.conductor.heimdall.core.entity.App;
 import br.com.conductor.heimdall.core.repository.AppRepository;
 import br.com.conductor.heimdall.core.util.Constants;
-import br.com.conductor.heimdall.gateway.trace.FilterDetail;
-import br.com.conductor.heimdall.gateway.trace.StackTraceImpl;
+import br.com.conductor.heimdall.core.trace.FilterDetail;
 import br.com.conductor.heimdall.gateway.trace.TraceContextHolder;
 import com.netflix.zuul.ZuulFilter;
 import com.netflix.zuul.context.RequestContext;
@@ -75,7 +74,7 @@ public class ScopesFilter extends ZuulFilter {
             detail.setStatus(Constants.SUCCESS);
         } catch (Throwable e) {
             detail.setStatus(Constants.FAILED);
-            TraceContextHolder.getInstance().getActualTrace().setStackTrace(new StackTraceImpl(e.getClass().getName(), e.getMessage()));
+            throw e;
         } finally {
             long endTime = System.currentTimeMillis();
