--- heimdall-gateway/src/main/java/br/com/conductor/heimdall/gateway/trace/RequestResponseParser.java
@@ -1,48 +1,0 @@
-
-package br.com.conductor.heimdall.gateway.trace;
-
-/*-
- * =========================LICENSE_START==================================
- * heimdall-gateway
- * ========================================================================
- * Copyright (C) 2018 Conductor Tecnologia SA
- * ========================================================================
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * ==========================LICENSE_END===================================
- */
-
-import java.util.Map;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonInclude.Include;
-import lombok.Data;
-
-/**
- * Data class a that represents a Request Response parser.
- *
- * @author Thiago Sampaio
- * @author Marcelo Aguiar Rodrigues
- */
-@Data
-public class RequestResponseParser {
-
-     @JsonInclude(Include.NON_NULL)
-     private String uri;
-
-     @JsonInclude(Include.NON_NULL)
-     private Map<String, String> headers;
-
-     @JsonInclude(Include.NON_NULL)
-     private String body;
-
-}
