+++ heimdall-gateway/src/test/java/br/com/conductor/heimdall/gateway/filter/ScopesFilterTest.java
@@ -22,7 +22,7 @@ package br.com.conductor.heimdall.gateway.filter;
 import br.com.conductor.heimdall.core.entity.*;
 import br.com.conductor.heimdall.core.repository.AppRepository;
 import br.com.conductor.heimdall.core.util.Constants;
-import br.com.conductor.heimdall.gateway.trace.FilterDetail;
+import br.com.conductor.heimdall.core.trace.FilterDetail;
 import br.com.conductor.heimdall.gateway.trace.TraceContextHolder;
 import com.netflix.zuul.context.RequestContext;
 import org.junit.Before;
